implement jack's suggestion of moving the ExtendedTextMapGetter from context.internal.propagation package into context.propagation.internal package
